fix deleting cart item in data base when removing item move isCartOwner method to new class to reuse it
