+++ src/main/java/com/gustavohenning/dbecommercev1/config/SecurityConfiguration.java
@@ -1,5 +1,8 @@
 package com.gustavohenning.dbecommercev1.config;
 
+import com.gustavohenning.dbecommercev1.repository.UserRepository;
+import com.gustavohenning.dbecommercev1.service.impl.CartServiceImpl;
+import com.gustavohenning.dbecommercev1.util.CartOwner;
 import com.gustavohenning.dbecommercev1.util.ExtractUserFromToken;
 import com.gustavohenning.dbecommercev1.util.RSAKeyProperties;
 import com.nimbusds.jose.jwk.JWK;
@@ -95,6 +98,11 @@ public class SecurityConfiguration {
     @Bean
     public ExtractUserFromToken extractUserFromToken(JwtDecoder jwtDecoder) {
         return new ExtractUserFromToken(jwtDecoder);
+    }
+
+    @Bean
+    public CartOwner cartOwner(UserRepository userRepository, CartServiceImpl cartService) {
+        return new CartOwner(userRepository, cartService);
     }
 
     @Bean
